Added ToString to all nodes

+++ src/HotChocolate/Language/src/Language.SyntaxTree/FloatFormat.cs
@@ -1,6 +1,6 @@
 namespace HotChocolate.Language
 {
-    public enum FloatFormat : byte
+    public enum FloatFormat : int
     {
         FixedPoint = 0,
         Exponential = 1
